fix: generate `__VLS_StyleModules` at template body
