+++ test-workspace/tsc/passedFixtures/vue3/#4788/main.vue
@@ -1,0 +1,20 @@
+<script lang="ts">
+import { exactType } from '../../../passedFixtures/shared';
+
+export default {
+    setup() {
+        return {
+            exactType,
+            a: 1
+        }
+    }
+};
+</script>
+
+<template>
+    {{ exactType(a, 1 as number) }}
+</template>
+
+<style module>
+.foo { }
+</style>
